+++ R/firstlib.R
@@ -1,4 +1,4 @@
 .onAttach <- function(lib, pkg)
 {
-  packageStartupMessage("Sotkanet R tools (C) 2013-2021 rOpenGov\nhttp://ropengov.org")
+  packageStartupMessage("Sotkanet R tools (C) 2013-2022 rOpenGov\nhttp://ropengov.org")
 }