+++ R/json.R
@@ -3,31 +3,19 @@
 #' @description Retrieve data from the query url.
 #' @param url Sotkanet JSON url
 #' @return sotkanet json query
-#' @importFrom rjson fromJSON
 #' @references See citation("sotkanet") 
 #' @author Maintainer: Leo Lahti \email{leo.lahti@@iki.fi}
+#' @importFrom jsonlite fromJSON
 #' @keywords utilities
-sotkanet.json_query <- function(url)
+sotkanet.json_query <- function(url, flatten = TRUE, ...)
 {
 
-  # Check that the URL exists
-  conn<-url(url)
-  doesnotexist<-inherits(try(suppressWarnings(readLines(conn)),silent=TRUE),"try-error")
-  close(conn)
-  if (doesnotexist) {
-    warning(paste("Sotkanet URL", url, "does not exist - returning NULL!"))
-    return(NULL)
+  # Check that URL fulfills requirements
+  # If not, test_connection returns a message and NULL
+  if (is.null(test_connection(url, ...))) {
+    return(invisible(NULL))
   }
 
-  # Retrieve the data
-  con <- url(url, method = "libcurl")
-  txt <- suppressWarnings(readLines(con, warn = FALSE))
-  close(con)
-  
-  response <- fromJSON(paste(txt, collapse = ""))
-  
-  if (is.null(response)) 
-    stop("Sotkanet server is not responding! Unable to query!")
-
-  return(response)
+  response <- jsonlite::fromJSON(url, flatten)
+  response
 }