+++ R/sotkanet-package.R
@@ -1,11 +1,11 @@
 #' @title Tools for Sotkanet API
-#' @description Retrieval of sotkanet data.
+#' @description Retrieval of sotkanet data: Statistical information on welfare and health in Finland.
 #'
 #' \tabular{ll}{
 #'   Package: \tab sotkanet\cr
 #'   Type: \tab Package\cr
 #'   Version: \tab See sessionInfo() or DESCRIPTION file\cr
-#'   Date: \tab 2013-2021\cr
+#'   Date: \tab 2013-2022\cr
 #'   License: \tab FreeBSD\cr
 #'   LazyLoad: \tab yes\cr
 #' }
@@ -13,7 +13,7 @@
 #' @name sotkanet-package
 #' @aliases sotkanet
 #' @docType package
-#' @author Leo Lahti, Einari Happonen, Juuso Parkkinen Joona Lehtomaki, Vesa Saaristo and Pyry Kantanen \email{leo.lahti@@iki.fi}
+#' @author Leo Lahti (\email{leo.lahti@@iki.fi}), Einari Happonen, Juuso Parkkinen Joona Lehtomaki, Vesa Saaristo and Pyry Kantanen
 #' @references See citation("sotkanet")
 #' @examples library(sotkanet)
 #' @keywords package