+++ NEWS.md
@@ -1,3 +1,20 @@
+# sotkanet 0.9.76 (2022-01-31)
+
+### NEW FEATURES
+ * Functions pointing to external resources (`sotkanet.json_query` and `sotkanet.csv_query`) should now fail gracefully. New internal functions are in graceful_http.R file
+ * New dependencies / imports: httr package for handling URLs and error messages and curl package for internet connection testing. jsonlite for handling data from json files
+ 
+### DEPRECATED AND DEFUNCT
+
+ * Removed internal functions: `GetDataSotkanetSingleIndicator` and `SotkanetData`.
+ * Removed dependency / import: rjson
+ 
+### MINOR IMPROVEMENTS
+
+ * Functionalities from GetDataSotkanetSingleIndicator and SotkanetData are now embedded in one function, `GetDataSotkanet`.
+ * Some files in ./R/ folder have been renamed for better legibility: internal.R to SotkanetCollect.R and sotkanet.R to SotkanetRegions.R
+ * New alias for downloading indicator metadata: `SotkanetIndicatorMetadata()` (usesSotkanetIndicator(id))
+
 # sotkanet 0.9.75 (2021-04-29)
 
  * Adding rmarkdown to Suggests